Add files via upload
To run the code, download the files "logo.JPG" and "PhoneOrder.py" into the same folder. Go to the folder and copy the full local address of the logo file (e.g. C:\\Users\\impli\\OneDrive\\Desktop\\python\\logo.jpg) and replace the logo address with it (highlighted in the picture).